requirements.txt
babel==2.17.0
customtkinter==5.2.2
darkdetect==0.8.0
decorator==5.2.1
imageio==2.37.0
imageio-ffmpeg==0.6.0
moviepy==2.1.2
numpy==2.2.4
packaging==24.2
pillow==10.4.0
proglog==0.1.10
pygame==2.6.1
python-dotenv==1.0.1
tkcalendar==1.6.1
tqdm==4.67.1
